Refactor the scheduler interface
