--- libcaf_core/caf/policy/unprofiled.cpp
@@ -1,13 +1,0 @@
-// This file is part of CAF, the C++ Actor Framework. See the file LICENSE in
-// the main distribution directory for license terms and copyright or visit
-// https://github.com/actor-framework/actor-framework/blob/master/LICENSE.
-
-#include "caf/policy/unprofiled.hpp"
-
-namespace caf::policy {
-
-unprofiled::~unprofiled() {
-  // nop
-}
-
-} // namespace caf::policy
