--- libcaf_core/caf/policy/work_sharing.cpp
@@ -1,13 +1,0 @@
-// This file is part of CAF, the C++ Actor Framework. See the file LICENSE in
-// the main distribution directory for license terms and copyright or visit
-// https://github.com/actor-framework/actor-framework/blob/master/LICENSE.
-
-#include "caf/policy/work_sharing.hpp"
-
-namespace caf::policy {
-
-work_sharing::~work_sharing() {
-  // nop
-}
-
-} // namespace caf::policy
